Small tweaks, enemy experimentation.
