+++ MarballsUnity/Assets/Scripts/Player Scripts/FollowObject.cs
@@ -2,7 +2,7 @@
 /// FollowObject.cs
 /// Authors: Kyle Dawson
 /// Date Created:  Feb. 18, 2015
-/// Last Revision: Apr. 18, 2015
+/// Last Revision: Apr. 22, 2015
 /// 
 /// Class that simply follows a specified object.
 /// 
@@ -13,8 +13,10 @@ using System.Collections;
 
 public class FollowObject : MonoBehaviour {
 	
-	protected Transform myTransform;	// Reference to own transform for caching.
-	public Transform trackingObject;	// Reference to transform of the position to be tracked.
+	protected Transform myTransform;		// Reference to own transform for caching.
+	public Transform trackingObject;		// Reference to transform of the position to be tracked.
+	public bool copyRotation = false;		// Whether this object should copy the rotation as well.
+	public Vector3 offset = Vector3.zero;	// How much distance there should be between object and follower.
 
 	// Start - Use this for initialization
 	protected virtual void Start () {
@@ -23,7 +25,11 @@ public class FollowObject : MonoBehaviour {
 	
 	// Update - Called once per frame.
 	protected virtual void Update () {
-		if (trackingObject != null)
-			myTransform.position = trackingObject.position;
+		if (trackingObject != null) {
+			myTransform.position = trackingObject.position + offset;
+
+			if (copyRotation)
+				myTransform.rotation = trackingObject.rotation;
+		}
 	}
 }
